Another update to the zip. The last?

+++ TelligentLocalInstance/Initialize.ps1
@@ -219,16 +219,17 @@ function Install-Solr {
     Start-Service tomcat* -ErrorAction Continue
 
 	# Install new Solr 6+ Version
-	$PackageName = '6-3-0.zip'
+    $PackageName = '6-3-0.zip'
     $FilePath = Join-Path $solrBase $PackageName
+    if(!(Test-Path $war) -or $Force) {
+        Invoke-WebRequest -Uri "https://github.com/afscrome/TelligentInstanceManager/raw/10updates/Solr/$($PackageName)?raw=true" -OutFile $FilePath
     
-    Invoke-WebRequest -Uri "https://github.com/afscrome/TelligentInstanceManager/raw/10updates/Solr/$($PackageName)?raw=true" -OutFile $FilePath
-    
-    #Expand-Zip $FilePath $SolrBase
-    Remove-Item $FilePath
+        #Expand-Zip $FilePath $SolrBase
+        Remove-Item $FilePath
 
-    $InstallScript = Join-Path $SolrBase 'Solr-6-3-0/bin/ServiceInstall.ps1'
-    &$InstallScript -ServiceName TIM-Search-6-3-0 -DisplayName TIM Search 6.3.0 -Port 8630
+        $InstallScript = Join-Path $SolrBase 'Solr-6-3-0/bin/ServiceInstall.ps1'
+        &$InstallScript -ServiceName "TIM-Search-6-3-0" -DisplayName "TIM Search 6.3.0" -Port 8630
+    }
 }
 
 function Initalize-Environment
